+++ client/Commands/Toolbar.test.tsx
@@ -2,7 +2,7 @@ import * as Enzyme from "enzyme";
 import * as Adapter from "enzyme-adapter-react-16";
 import * as React from "react";
 
-import { Command } from "../Command";
+import { Command } from "./Command";
 import { Toolbar } from "./Toolbar";
 
 Enzyme.configure({ adapter: new Adapter() });
