+++ client/Commands/Toolbar.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
-import { Button } from "../../Components/Button";
-import { Command } from "../Command";
+import { Button } from "../Components/Button";
+import { Command } from "./Command";
 
 interface ToolbarProps {
   encounterCommands: Command[];
